moved files, added spark lib info
