+++ spark.json
@@ -1,0 +1,7 @@
+{
+    "name": "ADAFRUIT_ST7565",
+    "version": "0.0.1",
+    "author": "Someone <email@somesite.com>",
+    "license": "Choose a license",
+    "description": "Briefly describe this library"
+}
